requirements.txt
streamlit
requests
fastapi
uvicorn
pypdf
langchain
sentence-transformers
faiss-cpu
python-multipart
groq